requirements.txt
numpy~=1.19.5
matplotlib~=3.4.2
pandas~=1.3.4
scipy~=1.7.0
tqdm~=4.64.1
PyYAML~=6.0
pip~=23.0
wheel~=0.36.2
setuptools~=57.0.0
torchvision~=0.14.1
torch~=1.13.1
Pillow~=9.4.0
nltk~=3.8.1
ipython~=8.11.0
scikit-learn~=0.24.2
mlinsights
